pyproject.toml
[tool.poetry]
name = "cuticulus"
version = "0.1.0"
description = ""
authors = ["Noah Gardner <ngngardner@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
jupyter = "^1.0.0"
ipykernel = "^6.9.1"

beartype = "^0.9.0"
gdown = "^4.2.0"
openpyxl = "^3.0.9"
python-dotenv = "^0.19.1"

questionary = "^1.10.0"
rich = "^10.12.0"

matplotlib = "^3.5.1"
numpy = "^1.22.2"
pandas = "^1.3.4"
Pillow = "^8.4.0"
scikit-learn = "0.24.1"
scipy = "1.6.0"
seaborn = "^0.11.2"
torch = "^1.10.2"

pygame = "^2.1.2"
grad-cam = "^1.3.7"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
flake8 = "^3.9.2"
wemake-python-styleguide = "^0.15.3"
pyright = "^0.0.10"
bandit = "1.7.2"

[tool.poetry.scripts]
cuticulus = 'cuticulus.__main__:main'
gui = 'gui.__main__:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
